+++ Sudoku/SudokuClient/Tools/TextBoxParametres.cs
@@ -6,15 +6,13 @@ namespace SudokuClient.Tools
 {
 	public class TextBoxParametres
 	{
-		public TextBoxParametres(int x, int y)
+		public TextBoxParametres(byte x, byte y)
 		{
 			X = x;
 			Y = y;
 		}
 
-		public int X { get; private set; }
-		public int Y { get; private set; }
-
-
+		public byte X { get; private set; }
+		public byte Y { get; private set; }
 	}
 }
